+++ CHANGELOG.md
@@ -3,6 +3,10 @@ SPDX-FileCopyrightText: 2020 Robin Krahl <robin.krahl@ireas.org>
 SPDX-License-Identifier: CC0-1.0
 -->
 
+# Unreleased
+
+- Add `Strikethrough` effect.
+
 # v0.2.0 (2020-10-03)
 
 This minor release refactors `Effects`, adds setters to `Style` and changes the